+++ cedar-policy-cli/CHANGELOG.md
@@ -7,7 +7,7 @@ Changes to the Cedar language, which are likely to affect users of the CLI, are 
 
 ### Changed
 
-- The `authorize` command will now error if used with the `--template-linked` argument will now error if the link file does not exist.
+- The `authorize` command with the `--template-linked` argument will now error if the link file does not exist.
 
 ## 4.11.1